feat: adding transactions list
