+++ lib/core/routes.dart
@@ -1,16 +1,16 @@
-import 'package:crypto_listing/convert/widgets/check_out_page.dart';
-import 'package:crypto_listing/convert/widgets/complete_confirmation.dart';
+import '../review/view/review_page.dart';
+import '../confirmation/view/complete_confirmation.dart';
 
 import '../convert/view/convert_screen.dart';
 import '../wallet/view/wallet_screen.dart';
 import '../cripto_details/view/detail_screen.dart';
-import '../cripto_movements/view/movements_screen.dart';
+import '../transactions/view/transactions_screen.dart';
 
 var routesNavigation = {
   '/wallet': (context) => const WalletScreen(),
-  '/criptoMoviments': (context) => const MovementsScreen(),
+  '/criptoMoviments': (context) => const TransactionsScreen(),
   '/criptoDetails': (context) => const DetailScreen(),
   '/convert': (context) => const ConvertScreen(),
-  '/confirm-conversion': (context) => const CheckOutPage(),
+  '/confirm-conversion': (context) => const ReviewPage(),
   '/complete-conversion': (context) => const CompleteConfirmation()
 };
