+++ lib/cripto_details/usecases/get_detail_usecase.dart
@@ -1,5 +1,4 @@
 import '../mapper/detail_mapper.dart';
-
 import '../model/detail_list_view_data.dart';
 import '../repository/get_detail_repo.dart';
 import '../repository/responses/detail_response.dart';
