+++ lib/shared/providers/providers.dart
@@ -1,7 +1,8 @@
-import 'total_amount_provider.dart';
 import 'package:decimal/decimal.dart';
 import 'package:dio/dio.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+import 'total_amount_provider.dart';
 
 final hideWalletStateProvider = StateProvider((ref) => false);
 final cryptoNameProvider = StateProvider((ref) => '');
