+++ lib/transactions/model/transaction_model.dart
@@ -1,0 +1,17 @@
+class TransactionModel {
+  DateTime date;
+  double fromCryptoQtd;
+  String fromCryptoAbrev;
+  double toCryptoQtd;
+  String toCryptoAbrev;
+  double valueReais;
+
+  TransactionModel({
+    required this.date,
+    required this.fromCryptoQtd,
+    required this.fromCryptoAbrev,
+    required this.toCryptoQtd,
+    required this.toCryptoAbrev,
+    required this.valueReais,
+  });
+}
