+++ lib/transactions/providers/transaction_providers.dart
@@ -1,0 +1,3 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+final transactionsProvider = StateProvider((ref) => []);
